QR_Backend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.1.7
django-auto-logout==0.5.1
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
html5lib==1.1
idna==3.10
lxml==5.3.1
oscrypto==1.3.0
pillow==11.1.0
pycparser==2.22
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.9.0
pypdf==5.4.0
python-bidi==0.6.6
PyYAML==6.0.2
qrcode==8.0
reportlab==4.3.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.13.0
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.3.0
webencodings==0.5.1
xhtml2pdf==0.2.17
